+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.6] - Unreleased
+
+### Added
+
+- Added rich.align.Align class
+- Added justify argument to Console.print and console.log
+
 ## [1.1.5] - 2020-05-15
 
 ### Changed